# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Incident {
  id: number;
  incidentNumber: string;
  title: string;
  description: string;
  category: string;
  severity: string;
  priority: string;
  riskScore: number;
  status: string;
  reporter: string;
  department: string;
  /** @nullable */
  assignedTo: string | null;
  /** @nullable */
  affectedSystem?: string | null;
  affectedUsers?: number;
  createdAt: string;
  updatedAt: string;
  slaDueAt?: string;
}

export interface TimelineEvent {
  id: number;
  label: string;
  detail: string;
  actor: string;
  kind: string;
  createdAt: string;
}

export type IncidentCommentVisibility = typeof IncidentCommentVisibility[keyof typeof IncidentCommentVisibility];


export const IncidentCommentVisibility = {
  reporter: 'reporter',
  internal: 'internal',
} as const;

export interface IncidentComment {
  id: number;
  author: string;
  message: string;
  visibility: IncidentCommentVisibility;
  createdAt: string;
}

export interface Evidence {
  id: number;
  filename: string;
  type: string;
  size: string;
  hash: string;
  uploadedBy: string;
  uploadedAt: string;
}

export interface Ioc {
  id: number;
  type: string;
  value: string;
  confidence: number;
}

export interface ChecklistItem {
  id: number;
  label: string;
  completed: boolean;
}

export type IncidentDetail = Incident & ({
  incidentDate?: string;
  discoveryDate?: string;
  /** @nullable */
  location?: string | null;
  impact?: string;
  /** @nullable */
  rootCause?: string | null;
  /** @nullable */
  attackVector?: string | null;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  containment?: string | null;
  /** @nullable */
  eradication?: string | null;
  /** @nullable */
  recovery?: string | null;
  timeline: TimelineEvent[];
  comments: IncidentComment[];
  evidence: Evidence[];
  iocs: Ioc[];
  checklist: ChecklistItem[];
});

export interface IncidentInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  category: string;
  reporter: string;
  department: string;
  severity?: string;
  affectedSystem?: string;
  /** @minimum 0 */
  affectedUsers?: number;
  impact?: string;
  location?: string;
}

export interface IncidentUpdate {
  status?: string;
  severity?: string;
  priority?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore?: number;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  rootCause?: string | null;
  /** @nullable */
  attackVector?: string | null;
  /** @nullable */
  containment?: string | null;
  /** @nullable */
  eradication?: string | null;
  /** @nullable */
  recovery?: string | null;
  /** @nullable */
  resolution?: string | null;
}

export type CommentInputVisibility = typeof CommentInputVisibility[keyof typeof CommentInputVisibility];


export const CommentInputVisibility = {
  reporter: 'reporter',
  internal: 'internal',
} as const;

export interface CommentInput {
  author: string;
  /** @minLength 1 */
  message: string;
  visibility: CommentInputVisibility;
}

export interface CountItem {
  name: string;
  count: number;
}

export interface TrendItem {
  day: string;
  reported: number;
  resolved: number;
}

export interface Dashboard {
  total: number;
  open: number;
  critical: number;
  highRisk: number;
  slaAtRisk: number;
  slaBreached: number;
  resolved: number;
  categories: CountItem[];
  severities: CountItem[];
  weeklyTrend: TrendItem[];
  recentActivity: TimelineEvent[];
}

export type ListIncidentsParams = {
search?: string;
status?: string;
severity?: string;
};