# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */

export interface ChecklistItem {
  id: number;
  label: string;
  completed: boolean;
}