# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */
import type { CommentInputVisibility } from './commentInputVisibility';

export interface CommentInput {
  author: string;
  /** @minLength 1 */
  message: string;
  visibility: CommentInputVisibility;
}