# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */
import type { CountItem } from './countItem';
import type { TimelineEvent } from './timelineEvent';
import type { TrendItem } from './trendItem';

export interface Dashboard {
  total: number;
  open: number;
  critical: number;
  highRisk: number;
  slaAtRisk: number;
  slaBreached: number;
  resolved: number;
  categories: CountItem[];
  severities: CountItem[];
  weeklyTrend: TrendItem[];
  recentActivity: TimelineEvent[];
}