# lib/api-zod/src/generated/types/evidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */

export interface Evidence {
  id: number;
  filename: string;
  type: string;
  size: string;
  hash: string;
  uploadedBy: string;
  uploadedAt: Date;
}