# lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */

export interface Incident {
  id: number;
  incidentNumber: string;
  title: string;
  description: string;
  category: string;
  severity: string;
  priority: string;
  riskScore: number;
  status: string;
  reporter: string;
  department: string;
  /** @nullable */
  assignedTo: string | null;
  /** @nullable */
  affectedSystem?: string | null;
  affectedUsers?: number;
  createdAt: Date;
  updatedAt: Date;
  slaDueAt?: Date;
}