# lib/api-zod/src/generated/types/incidentComment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentCommentVisibility } from './incidentCommentVisibility';

export interface IncidentComment {
  id: number;
  author: string;
  message: string;
  visibility: IncidentCommentVisibility;
  createdAt: Date;
}