# lib/api-zod/src/generated/types/incidentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberShield defensive incident management API
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  category: string;
  reporter: string;
  department: string;
  severity?: string;
  affectedSystem?: string;
  /** @minimum 0 */
  affectedUsers?: number;
  impact?: string;
  location?: string;
}